third_party/blink/web_tests/wpt_internal/webgpu/canvas_webgpu_transfer/transferBackFromGPUTexture-destroys-texture.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="./canvas-test.js"></script>
<script src="./webgpu-helpers.js"></script>

<h1>Canvas2D WebGPU Transfer: transferBackFromGPUTexture-destroys-texture</h1>
<p>transferBackFromGPUTexture() causes the GPUTexture returned by
transferToGPUTexture() to enter a destroyed state.</p>

<script>

runCanvasTestsInWorker({dependencies: ['./webgpu-helpers.js']});

/**
 * transferBackFromGPUTexture() should cause the GPUTexture returned by
 * transferToGPUTexture() to enter a destroyed state.
 */
canvasPromiseTest(async (canvas) => {
  const device = await getWebGPUDevice();

  // Begin and end a WebGPU access session.
  const ctx = canvas.getContext('2d');
  const tex = ctx.transferToGPUTexture({device: device,
                                        usage: GPUTextureUsage.COPY_SRC |
                                               GPUTextureUsage.COPY_DST});
  ctx.transferBackFromGPUTexture();

  // `tex` should be in a destroyed state.
  assert_true(await isTextureDestroyed(device, tex));
}, 'transferBackFromGPUTexture() should destroy the associated GPUTexture.');

</script>
